Add inline admin support!
Turns out to be really easy to implement,
as the TranslatableModelForm does all the form.
Moved the shared logic to an admin base class.

The current implementation does require the parent model to be a
TranslatableModel, as it's used to display the global language tabs.
The available languages of the parent model define which inlines are
displayed.